requirements.txt
spotipy
python-dotenv
argparse
